orchestration/requirements.txt
# LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.0.20

# LangFuse for observability
langfuse>=2.0.0

# HTTP clients
httpx>=0.26.0
requests>=2.31.0
sseclient-py>=1.8.0

# OpenTelemetry
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-jaeger>=1.21.0

# JWT handling
PyJWT[crypto]>=2.8.0
cryptography>=41.0.0

# Data handling
pydantic>=2.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
